refactor: 고객이 남기는 후기 별점 타입 변경 (int -> double)

+++ backend/src/main/java/com/yourseason/backend/member/customer/controller/dto/ReviewListResponse.java
@@ -13,7 +13,7 @@ public class ReviewListResponse {
     private Long reviewId;
     private String consultantNickname;
     private String consultantImageUrl;
-    private int star;
+    private double star;
     private String comment;
     private LocalDate reviewDate;
 }
